# Conflicts:
#	app/src/main/java/org/ole/planet/myplanet/data/room/AppDatabase.kt
#	app/src/main/java/org/ole/planet/myplanet/repository/EventsRepositoryImpl.kt
#	app/src/main/java/org/ole/planet/myplanet/services/TaskNotificationWorker.kt
#	app/src/main/res/values/strings.xml
#	app/src/test/java/org/ole/planet/myplanet/repository/EventsRepositoryImplTest.kt
#	app/src/test/java/org/ole/planet/myplanet/repository/TeamsRepositoryImplTest.kt
#	app/src/test/java/org/ole/planet/myplanet/services/TaskNotificationWorkerTest.kt

+++ app/src/main/java/org/ole/planet/myplanet/utils/MapTileUtils.kt
@@ -2,6 +2,7 @@ package org.ole.planet.myplanet.utils
 
 import android.content.Context
 import android.os.Environment
+import android.util.Log
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
@@ -12,17 +13,24 @@ object MapTileUtils {
     fun copyAssets(context: Context) {
         val tiles = arrayOf("dhulikhel.mbtiles", "somalia.mbtiles")
         val assetManager = context.assets
-        try {
-            for (s in tiles) {
+        for (s in tiles) {
+            try {
                 val outFile = File(Environment.getExternalStorageDirectory().toString() + "/osmdroid", s)
+
+                if (outFile.exists() && outFile.length() > 0) {
+                    continue
+                }
+
+                outFile.parentFile?.mkdirs()
+
                 assetManager.open(s).use { input ->
                     FileOutputStream(outFile).use { output ->
                         copyFile(input, output)
                     }
                 }
+            } catch (e: Exception) {
+                Log.w("MapTileUtils", "Failed to copy map tile: $s", e)
             }
-        } catch (e: Exception) {
-            e.printStackTrace()
         }
     }
